# expression: implement vectorized evaluation for builtinAddDateStringIntSig