@b41sh 2019-11-15
We use generator to generate some similar expressions,
for example this PR [13066](https://github.com/pingcap/tidb/pull/13066) generate all `builtin*TimeDiffSig`,
Are you interested to writing a generator to generate all `builtinAddDate*Sig`?
You can refer to this document https://docs.google.com/document/d/17XdYo_EwYaWANqbjS3i3OVSz3oLlovU2fIo1bRx4YqA/edit?pli=1#