@ekalinin 2019-11-21
> We use generator to generate some similar expressions,
for example this PR 13066 generate all builtin*TimeDiffSig,

Cool, will take a look.

> Are you interested to writing a generator to generate all builtinAddDate*Sig?

Yeah, will try to add it.